ingestCalibs: fix output directory for registry
We aim to have the calib registry end up in the calib root directory. To
this end, we modify RegisterTask.openRegistry to take a directory instead
of the butler (the butler was only being used to determine the directory).

Our only access to the calib root directory is through the "--calib"
command-line argument; if that is not provided, write to the current
working directory for the user to move into the correct place.